+++ .github/ISSUE_TEMPLATE/bug.yml
@@ -1,7 +1,7 @@
 name: ​🐞 Bug
 description: Report an issue to help us improve the project.
 title: '[BUG] '
-labels: ["bug"]
+labels: ['bug']
 body:
   - type: textarea
     attributes:
