+++ .github/workflows/auto-label.yml
@@ -18,7 +18,7 @@ jobs:
             const issue = context.payload.issue;
             const issueBody = issue.body ? issue.body.toLowerCase() : '';
             const issueTitle = issue.title.toLowerCase();
-            
+
             // Add gssoc label to all issues
             await github.rest.issues.addLabels({
               owner: context.repo.owner,
@@ -34,4 +34,3 @@ jobs:
                 labels: [label]
               });
             };
-
