+++ PULL_REQUEST_TEMPLATE.md
@@ -1,40 +1,47 @@
 # Description
 
 **Summary of Changes:**
+
 - Provide a brief overview of the changes made in relation to the issue.
 
 **Context:**
+
 - Explain the motivation behind your changes and any specific problems they address.
 
 **Tech Stack & Dependencies:**
+
 - List the technologies and any dependencies required for this change.
 
 ---
 
 ## Type of Change
+
 - [ ] Feature Added (Issue number: <!-- write issue number here -->)
 - [ ] Bug Fix (Issue number: <!-- write issue number here -->)
 <!-- Indicate the type of change by marking [X] in the appropriate box. Example: [X] Bug Fix -->
 
 ---
 
-## Screenshots 
+## Screenshots
 
 **Screenshots:**
+
 - Insert relevant screenshots here.
 
 ---
 
 # Checklist:
-- [X] I have included a `README.md` file that clearly explains the project's purpose, functionality, and usage instructions (if applicable to new features).
-- [X] My code adheres to the project's style guidelines.
-- [X] I have thoroughly reviewed my code and addressed potential errors or inconsistencies.
-- [X] I have added clear and concise comments to my code where necessary, particularly in complex sections.
-- [X] I have reviewed and followed the project's [Contributing Guidelines](CONTRIBUTING.md) to ensure my contribution aligns with the project's requirements.
+
+- [x] I have included a `README.md` file that clearly explains the project's purpose, functionality, and usage instructions (if applicable to new features).
+- [x] My code adheres to the project's style guidelines.
+- [x] I have thoroughly reviewed my code and addressed potential errors or inconsistencies.
+- [x] I have added clear and concise comments to my code where necessary, particularly in complex sections.
+- [x] I have reviewed and followed the project's [Contributing Guidelines](CONTRIBUTING.md) to ensure my contribution aligns with the project's requirements.
 
 ---
 
 ### Additional Notes
+
 - Add any additional comments or insights that may assist in reviewing your PR.
 
 <!-- Write all the context clearly and concisely. Remove instruction lines before submitting your PR. -->
