+++ eslint.config.js
@@ -1,0 +1,10 @@
+const config = [
+  {
+    files: ['*.js', '*.ejs', '*.html'], // Specify file types to lint
+    rules: {
+      // Place any custom rules here
+    },
+  },
+];
+
+module.exports = config;
