+++ package.json
@@ -4,7 +4,10 @@
   "main": "index.js",
   "scripts": {
     "test": "npx nodemon app.js",
-    "start": "node app.js"
+    "start": "node app.js",
+    "lint": "eslint .",
+    "lint:ejs": "npx ejs-lint views/**/*.ejs",
+    "format": "prettier --write ."
   },
   "author": "",
   "license": "ISC",
@@ -14,6 +17,7 @@
     "connect-mongo": "^5.1.0",
     "dotenv": "^16.4.5",
     "ejs": "^3.1.10",
+    "ejs-lint": "^2.0.0",
     "express": "^4.21.1",
     "express-session": "^1.18.1",
     "express-validator": "^7.1.0",
@@ -25,5 +29,12 @@
     "scruter": "file:"
   },
   "description": "Scruter is a local classifieds and community site designed to bring people together for buying, selling, and exchanging goods and services.",
-  "keywords": []
+  "keywords": [],
+  "devDependencies": {
+    "eslint": "^9.12.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-ejs": "^0.0.2",
+    "eslint-plugin-html": "^8.1.2",
+    "prettier": "^3.3.3"
+  }
 }
