+++ views/auth.ejs
@@ -3,8 +3,8 @@
 <head>
     <%- include('partials/head') %> 
     <link href='https://unpkg.com/boxicons@2.1.4/css/boxicons.min.css' rel='stylesheet'>
-    <link rel="stylesheet" href="formstyle.css">
-    <script src="/client-validation.js" defer></script>
+    <link rel="stylesheet" href="css/formstyle.css">
+    <script src="js/client-validation.js" defer></script>
     <style>
         body{
             overflow-y: hidden;
