+++ views/dashboard.ejs
@@ -1,13 +1,7 @@
 <!DOCTYPE html> 
 <html lang="en"> 
 <head> 
-    <title>User Dashboard</title>
     <%- include('partials/head') %> 
-    <style>
-        strong{
-            font-weight: bold;
-        }
-    </style>
 </head> 
 <body> 
     <%- include('partials/navbar') %>
