+++ views/form.ejs
@@ -101,8 +101,8 @@
             </div>
         </div>
     </section>
-    <script src="/search-auto.js"></script>
-    <script src="/client-validation.js"></script>
+    <script src="js/search-auto.js"></script>
+    <script src="js/client-validation.js"></script>
     <script>
         // /preloader js styling
         let preloader = document.querySelector("#preloader");
