+++ views/partials/footer.ejs
@@ -1,143 +1,7 @@
-<!-- Footer Section -->
-<style>
-  .social-section {
-    padding: 30px;
-  }
-  /* Media query for phone screens or smaller */
-  @media (max-width: 900px) {
-    body {
-      padding-bottom: 61px;
-    }
-  }
-
-  /* General styles for the footer */
-  .social-section {
-    background-color: #f8f9fa;
-    padding: 40px 0;
-    border-top: 1px solid #e0e0e0;
-  }
-
-  .social-section h4 {
-    font-size: 1.4rem;
-    font-weight: bold;
-    margin-bottom: 20px;
-  }
-
-  .social-section p {
-    font-size: 1.1rem;
-    color: #666;
-    max-width: 600px;
-    margin: 0 auto;
-  }
-
-  /* Adds a hover effect that smoothly scales up the text and underlines it,
-   providing a more interactive and visually appealing user experience */
-  .quick-link {
-    display: inline-block;
-    transition: transform 0.3s ease;
-  }
-
-  .quick-link:hover {
-    transform: scale(1.1);
-    text-decoration: underline;
-  }
-
-  /* Adds a hover effect that smoothly scales up the text and underlines it,
-   providing a more interactive and visually appealing user experience */
-  .help-desk {
-    display: inline-block;
-    transition: transform 0.3s ease;
-  }
-
-  .help-desk:hover {
-    transform: scale(1.1);
-    text-decoration: underline;
-  }
-
-  /* Styling for the links in the footer */
-  .social-section ul {
-    padding-left: 0;
-    list-style: none;
-  }
-
-  .social-section ul li {
-    margin-bottom: 10px;
-  }
-
-  .social-section ul li a {
-    text-decoration: none;
-    color: #333;
-    transition: color 0.3s ease;
-  }
-
-  .social-section ul li a:hover {
-    color: #007bff;
-  }
-
-  /* Styles for the social media icons */
-  .icon-circle {
-    width: 60px;
-    height: 60px;
-    border-radius: 50%;
-    background-color: #fff;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    transition: transform 0.3s ease, box-shadow 0.3s ease;
-    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
-    margin-bottom: 20px;
-  }
-
-  .icon-circle:hover {
-    transform: translateY(-5px);
-    box-shadow: 0 6px 15px rgba(0, 0, 0, 0.2);
-  }
-
-  .icon-circle a {
-    color: #333;
-    font-size: 1.8rem;
-  }
-
-  .icon-circle a:hover {
-    color: #007bff;
-  }
-
-  /* Gap between the social media icons */
-  .d-flex {
-    gap: 15px;
-  }
-
-  /* Copyright text styling */
-  .copyright-text {
-    margin-top: 40px;
-    font-size: 0.9rem;
-    color: #666;
-  }
-
-  /* Responsive design: For smaller screens */
-  @media (max-width: 768px) {
-    .social-section h4 {
-      font-size: 1.2rem;
-    }
-
-    .social-section ul li {
-      font-size: 1rem;
-    }
-
-    .icon-circle {
-      width: 50px;
-      height: 50px;
-    }
-
-    .icon-circle a {
-      font-size: 1.5rem;
-    }
-  }
-</style>
-<section class="text-center social-section bg-light py-5">
+<section id="footer" class="social-section py-5">
   <div class="container">
     <div class="row">
-      <!-- First part: Website description -->
+      <!-- About Scruter -->
       <div class="col-md-3">
         <h4>About Scruter</h4>
         <p>
@@ -147,29 +11,19 @@
         </p>
       </div>
 
-      <!-- Second part: Links to About and other pages -->
+      <!-- Quick Links -->
       <div class="col-md-3">
         <h4>Quick Links</h4>
         <ul class="list-unstyled">
           <li><a href="/about" class="text-dark quick-link">About Us</a></li>
           <li><a href="/team" class="text-dark quick-link">Our Team</a></li>
-          <li>
-            <a href="/privacy-policy" class="text-dark quick-link"
-              >Privacy Policy</a
-            >
-          </li>
-          <li>
-            <a href="/terms" class="text-dark quick-link">Terms of Service</a>
-          </li>
-          <li>
-            <a href="/contributors" class="text-dark quick-link"
-              >Our Contributors</a
-            >
-          </li>
+          <li><a href="/privacy-policy" class="text-dark quick-link">Privacy Policy</a></li>
+          <li><a href="/terms" class="text-dark quick-link">Terms of Service</a></li>
+          <li><a href="/contributors" class="text-dark quick-link">Our Contributors</a></li>
         </ul>
       </div>
 
-      <!-- Third part: Helpdesk, Queries, and Contact Us -->
+      <!-- Helpdesk -->
       <div class="col-md-3">
         <h4>Helpdesk</h4>
         <ul class="list-unstyled">
@@ -180,37 +34,22 @@
         </ul>
       </div>
 
-      <!-- Fourth part: Socials -->
+      <!-- Social Media Links -->
       <div class="col-md-3">
         <h4>Follow Us</h4>
         <div class="d-flex justify-content-center">
           <div class="icon-circle mx-2">
-            <a
-              href="https://facebook.com"
-              target="_blank"
-              aria-label="Follow us on Facebook"
-              class="text-dark"
-            >
+            <a href="https://facebook.com" target="_blank" aria-label="Follow us on Facebook">
               <i class="fa-brands fa-facebook fa-lg"></i>
             </a>
           </div>
           <div class="icon-circle mx-2">
-            <a
-              href="https://twitter.com"
-              target="_blank"
-              aria-label="Follow us on Twitter"
-              class="text-dark"
-            >
+            <a href="https://twitter.com" target="_blank" aria-label="Follow us on Twitter">
               <i class="fa-brands fa-x-twitter"></i>
             </a>
           </div>
           <div class="icon-circle mx-2">
-            <a
-              href="https://instagram.com"
-              target="_blank"
-              aria-label="Follow us on Instagram"
-              class="text-dark"
-            >
+            <a href="https://instagram.com" target="_blank" aria-label="Follow us on Instagram">
               <i class="fa-brands fa-instagram fa-lg"></i>
             </a>
           </div>
@@ -226,3 +65,25 @@
     </div>
   </div>
 </section>
+
+<script>
+  document.addEventListener('DOMContentLoaded', function () {
+    const toggleButton = document.getElementById('dark-mode-toggle');
+
+    // Load the user's preference from local storage
+    if (localStorage.getItem('dark-mode') === 'enabled') {
+      document.body.classList.add('dark-mode');
+    }
+
+    toggleButton.addEventListener('click', () => {
+      document.body.classList.toggle('dark-mode');
+
+      // Save the user's preference
+      if (document.body.classList.contains('dark-mode')) {
+        localStorage.setItem('dark-mode', 'enabled');
+      } else {
+        localStorage.removeItem('dark-mode');
+      }
+    });
+  });
+</script>
