+++ views/partials/searchbar.ejs
@@ -48,4 +48,3 @@
     });
   });
 </script>
-<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
